Re-organise with new architecture
