+++ src/lib/color-theme/interface/monacoTheme.interface.ts
