+++ src/lib/color-theme/interface/rule.interface.ts
