+++ src/lib/color-theme/interface/variables.interface.ts
