+++ src/lib/com/event/action-monaco-ready.interface.ts
