+++ src/lib/com/event/action-reset-theme.interface.ts
