+++ src/lib/com/event/base-action.interface.ts
