+++ src/lib/feature-detection/environmentDetection.ts
