+++ src/lib/feature-detection/event/ide-dom-hidden.event.ts
