+++ src/lib/feature-detection/event/ide-dom-updated.event.ts
