+++ src/lib/feature-detection/event/index.ts
