+++ src/lib/feature-detection/ideDomWatcher.ts
