+++ src/lib/feature-detection/ideVersion.enum.ts
