+++ src/lib/feature-detection/index.ts
