+++ src/lib/feature-detection/navigationDetection.ts
