+++ src/lib/folders/class/gasFile.ts
