+++ src/lib/folders/class/gasRoot.ts
