+++ src/lib/folders/folderState.interface.ts
